revert(activity): keep money-account activity types in shared
Restore shared activity types and revert UI imports to shared/lib/activity/types
after review feedback to remove the ui/selectors/activity/types split.
