+++ ui/pages/activity/useLocalTransactions.test.ts
@@ -1,6 +1,6 @@
 import { renderHook } from '@testing-library/react';
 import { useSelector } from 'react-redux';
-import type { ActivityListItem } from '../../selectors/activity/types';
+import type { ActivityListItem } from '../../../shared/lib/activity/types';
 import { useLocalTransactions } from './useLocalTransactions';
 
 jest.mock('react-redux', () => ({
