+++ ui/pages/details/components/sections.tsx
@@ -1,6 +1,8 @@
 import React from 'react';
-import type { ActivityListItem } from '../../../selectors/activity/types';
-import type { TokenAmount } from '../../../../shared/lib/activity/types';
+import type {
+  ActivityListItem,
+  TokenAmount,
+} from '../../../../shared/lib/activity/types';
 import { isValidTransactionHash } from '../../../../shared/lib/transactions.utils';
 import { AccountName } from '../../../components/app/transaction/account-name';
 import { NetworkName } from '../../../components/app/transaction/network-name';
