+++ ui/pages/details/templates/perps-details.tsx
@@ -1,7 +1,9 @@
 import React, { useMemo } from 'react';
 import { toEvmCaipChainId } from '@metamask/multichain-network-controller';
-import type { ActivityListItem } from '../../../selectors/activity/types';
-import type { TokenAmount } from '../../../../shared/lib/activity/types';
+import type {
+  ActivityListItem,
+  TokenAmount,
+} from '../../../../shared/lib/activity/types';
 import { toAssetId } from '../../../../shared/lib/asset-utils';
 import { AccountName } from '../../../components/app/transaction/account-name';
 import { NetworkName } from '../../../components/app/transaction/network-name';
