+++ ui/pages/details/templates/ramps/hooks.ts
@@ -1,5 +1,5 @@
 import { useMemo } from 'react';
-import type { ActivityListItem } from '../../../../selectors/activity/types';
+import type { ActivityListItem } from '../../../../../shared/lib/activity/types';
 import { useRampsOrders } from '../../../../hooks/ramps/useRampsOrders';
 import { mapRampsOrderSafely } from '../../../../hooks/ramps/utils/mapRampsOrderSafely';
