+++ ui/pages/details/templates/ramps/ramp-metadata-section.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import type { ActivityListItem } from '../../../../selectors/activity/types';
+import type { ActivityListItem } from '../../../../../shared/lib/activity/types';
 import { isValidTransactionHash } from '../../../../../shared/lib/transactions.utils';
 import { AccountName } from '../../../../components/app/transaction/account-name';
 import { NetworkName } from '../../../../components/app/transaction/network-name';
