+++ ui/selectors/activity/enrich-local-musd-claim.test.ts
@@ -6,6 +6,7 @@ import {
 } from '@metamask/transaction-controller';
 import { mapLocalTransaction } from '@metamask/client-utils';
 import { CHAIN_IDS } from '../../../shared/constants/network';
+import type { ActivityListItem } from '../../../shared/lib/activity/types';
 import { toAssetId } from '../../../shared/lib/asset-utils';
 import type { TransactionGroup } from '../../../shared/lib/multichain/types';
 import {
@@ -13,7 +14,6 @@ import {
   MERKL_DISTRIBUTOR_ADDRESS,
   MUSD_TOKEN_ADDRESS,
 } from '../../components/app/musd/constants';
-import type { ActivityListItem } from './types';
 import {
   enrichLocalMusdClaimActivity,
   resolveMusdClaimAmount,
